Backend/requirements.txt
fastapi==0.109.0
uvicorn==0.27.0
pydantic==2.6.1
pymongo==4.6.1
bcrypt==4.0.1
python-jose==3.3.0
python-multipart==0.0.6
opencv-python==4.9.0.80
numpy==1.26.4
face-recognition==1.3.0
requests==2.31.0
